fix DDL, add tests

+++ modules/api/src/main/java/org/apache/ignite/sql/async/AsyncResultSet.java
@@ -58,7 +58,7 @@ public interface AsyncResultSet {
      * @return Number of rows affected by the query, or {@code 0} if statement return nothing, or {@code -1} if inapplicable.
      * @see ResultSet#affectedRows()
      */
-    int affectedRows();
+    long affectedRows();
 
     /**
      * Returns whether the query that produce this result was a conditional query, or not. E.g. for the query "Create table if not exists"
